abstracted remaining api
