+++ mintapi/api.py
@@ -1,3 +1,5 @@
+from typing import Dict, List, Optional, Union
+
 from mintapi.browser import SeleniumBrowser
 from mintapi.rest import RESTClient
 
@@ -7,18 +9,29 @@ class Mint(object):
     Composed API client to route through the browser or REST calls
     """
 
-    def __init__(self, **kwargs):
+    def __init__(
+        self,
+        api_key: Optional[str] = None,
+        cookies: Optional[Union[str, List[Dict]]] = None,
+        **browser_params
+    ):
         """
         Passes forward parameters to the browser and rest client
 
         Pass a driver or email + password to authenticate with the browser
         OR pass in an api_key or cookie to auth directly with the rest client
-        """
-        self.browser = SeleniumBrowser(**kwargs)
-        self.rest_client = RESTClient(**kwargs)
 
-        if self.browser.driver is not None:
-            self.transfer_auth()
+        Browser is mainly used to generate auth (only necessary if not otherwise passed)
+        """
+        self.rest_client = RESTClient(api_key=api_key, cookies=cookies)
+
+        # only use browser if not sufficiently authorized already
+        if not api_key or not cookies:
+            self.browser = SeleniumBrowser(**browser_params)
+
+            if self.browser.driver is not None:
+                self.transfer_auth()
+            self.browser.close()
 
     def transfer_auth(self):
         api_key = self.browser._get_api_key_header()["authorization"]
