+++ mintapi/endpoints.py
@@ -5,11 +5,13 @@ Shared Endpoint module to keep definitions independent of implementation
 
 
 from abc import ABCMeta, abstractmethod
+from datetime import datetime
 from typing import List, Optional
 
 import pandas as pd
+from requests import Response
 
-from mintapi.constants import MINT_ROOT_URL
+from mintapi.constants import MINT_CREDIT_URL, MINT_ROOT_URL
 from mintapi.trends import (
     CategoryMatchFilter,
     DateFilter,
@@ -31,19 +33,101 @@ class MintEndpoints(object, metaclass=ABCMeta):
     def request(self):
         pass
 
-    @abstractmethod
-    def get(self):
-        pass
+    def get(self, **kwargs):
+        return self.request(method="GET", **kwargs)
 
-    @abstractmethod
-    def post(self):
-        pass
+    def post(self, **kwargs):
+        return self.request(method="POST", **kwargs)
+
+    def _paginate(self, data_key: str, metadata_key: str, response: Response, **kwargs):
+        """
+        Mint API appears to use a limit-offset pagination mechanism with
+        href links embedded in responses. Can iterate through sequentially
+        and append results together
+
+        Data schema:
+        {DataType: [], metaData: {}} where DataType is a dynamic key mapping to the endpoint
+        metaData follows a consistent format:
+
+        {
+            'asOf': 'TIMESTAMP',
+            'totalSize': INT,
+            'pageSize': 20,
+            'currentPage': 1,
+            'offset': 0,
+            'limit': 20,
+            'link': [
+                {'otherAttributes': {'method': 'GET'}, 'href': '/v1/accounts', 'rel': 'self'},
+                {'otherAttributes': {}, 'href': '/v1/accounts?offset=20&limit=20', 'rel': 'next'}
+            ]
+        }
+
+        Parameters
+        ----------
+        data_key : str
+            _description_
+        response : Response
+            _description_
+
+        Returns
+        -------
+        _type_
+            _description_
+        """
+        data = []
+
+        json_data = response.json()
+
+        # early abort if no data extraction theme
+        if data_key is None:
+            return json_data
+
+        data.extend(json_data[data_key])
+
+        # early abort if no pagination mechanism defined
+        if metadata_key is None:
+            return data
+
+        metadata = _ResponseMetadata(json_data[metadata_key])
+
+        # drop url params from propagating (href includes the full uri path already)
+        kwargs.pop("params", None)
+
+        while metadata.has_next:
+            response = self.request(
+                uri_path=metadata.next_uri_path,
+                data_key=data_key,
+                metadata_key=metadata_key,
+                paginate=False,
+                **kwargs,
+            )
+            json_data = response.json()
+            data.extend(json_data[data_key])
+            metadata = _ResponseMetadata(json_data[metadata_key])
+
+        return data
 
     """
     Endpoints - Acts as the api descriptor (equivalent to an openapi generated client)
     Separates out endpoint info to make future updates easy without changing the publicly
     exposed methods
     """
+
+    def _initiate_account_refresh(self, **kwargs):
+        api_url = MINT_ROOT_URL
+        api_section = ""
+        uri_path = "/refreshFILogins.xevent"
+        metadata_key = None
+        data_key = None
+
+        return self.post(
+            api_url=api_url,
+            api_section=api_section,
+            uri_path=uri_path,
+            data_key=data_key,
+            metadata_key=metadata_key,
+            **kwargs,
+        )
 
     def _get_account_data(self, **kwargs):
         api_url = MINT_ROOT_URL
@@ -58,7 +142,7 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     def _get_budget_data(self, **kwargs):
@@ -74,7 +158,7 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     def _get_bills_data(self, **kwargs):
@@ -90,7 +174,7 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     def _get_category_data(self, **kwargs):
@@ -106,7 +190,71 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
+        )
+
+    def _get_credit_accounts(self, **kwargs):
+        api_url = MINT_CREDIT_URL
+        api_section = ""
+        uri_path = "/v1/creditreports/0/tradelines"
+        metadata_key = None
+        data_key = None
+
+        return self.get(
+            api_url=api_url,
+            api_section=api_section,
+            uri_path=uri_path,
+            data_key=data_key,
+            metadata_key=metadata_key,
+            **kwargs,
+        )
+
+    def _get_credit_inquiries(self, **kwargs):
+        api_url = MINT_CREDIT_URL
+        api_section = ""
+        uri_path = "/v1/creditreports/0/inquiries"
+        metadata_key = None
+        data_key = None
+
+        return self.get(
+            api_url=api_url,
+            api_section=api_section,
+            uri_path=uri_path,
+            data_key=data_key,
+            metadata_key=metadata_key,
+            **kwargs,
+        )
+
+    def _get_credit_reports(self, **kwargs):
+        api_url = MINT_CREDIT_URL
+        api_section = ""
+        uri_path = "/v1/creditreports"
+        metadata_key = None
+        data_key = None
+
+        return self.get(
+            api_url=api_url,
+            api_section=api_section,
+            uri_path=uri_path,
+            data_key=data_key,
+            metadata_key=metadata_key,
+            **kwargs,
+        )
+
+    def _get_credit_utilization(self, **kwargs):
+        api_url = MINT_CREDIT_URL
+        api_section = ""
+        uri_path = "/v1/creditreports/creditutilizationhistory"
+        metadata_key = None
+        data_key = None
+
+        return self.get(
+            api_url=api_url,
+            api_section=api_section,
+            uri_path=uri_path,
+            data_key=data_key,
+            metadata_key=metadata_key,
+            **kwargs,
         )
 
     def _get_investment_data(self, **kwargs):
@@ -122,7 +270,7 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     def _get_transaction_data(self, **kwargs):
@@ -138,7 +286,7 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     def _get_trend_data(self, **kwargs):
@@ -154,12 +302,19 @@ class MintEndpoints(object, metaclass=ABCMeta):
             uri_path=uri_path,
             data_key=data_key,
             metadata_key=metadata_key,
-            **kwargs
+            **kwargs,
         )
 
     """
     User Methods - Add custom postprocessing logic here
     """
+
+    def initiate_account_refresh(self):
+        """
+        _summary_
+        """
+
+        self._initiate_account_refresh()
 
     def get_account_data(self, limit: int = 1000, **kwargs):
         """
@@ -239,6 +394,89 @@ class MintEndpoints(object, metaclass=ABCMeta):
         }
         params = {k: v for k, v in params.items() if v is not None}
         return self._get_category_data(params=params, **kwargs)
+
+    def get_credit_accounts(self, **kwargs):
+        """
+        _summary_
+
+        Returns
+        -------
+        _type_
+            _description_
+        """
+        return self._get_credit_accounts(**kwargs)
+
+    def get_credit_inquiries(self, **kwargs):
+        """
+        _summary_
+
+        Returns
+        -------
+        _type_
+            _description_
+        """
+        return self._get_credit_inquiries(**kwargs)
+
+    def get_credit_reports(self, limit: int = 1000, **kwargs):
+        """
+        _summary_
+
+        Parameters
+        ----------
+        limit : int, optional
+            _description_, by default 1000
+
+        Returns
+        -------
+        _type_
+            _description_
+        """
+        params = {
+            "limit": limit,
+        }
+        params = {k: v for k, v in params.items() if v is not None}
+        return self._get_credit_reports(params=params, **kwargs)
+
+    def get_credit_utilization(self, **kwargs):
+        """
+        _summary_
+
+        Returns
+        -------
+        _type_
+            _description_
+        """
+
+        def _process_utilization(data):
+            # Function to clean up the credit utilization history data
+            utilization = []
+            utilization.extend(_flatten_utilization(data["cumulative"]))
+            for trade in data["tradelines"]:
+                utilization.extend(_flatten_utilization(trade))
+            return utilization
+
+        def _flatten_utilization(data):
+            # The utilization history data has a nested format, grouped by year
+            # and then by month. Let's flatten that into a list of dates.
+            utilization = []
+            name = data.get("creditorName", "Total")
+            for cu in data["creditUtilization"]:
+                year = cu["year"]
+                for cu_month in cu["months"]:
+                    date = datetime.strptime(cu_month["name"], "%B").replace(
+                        day=1, year=int(year)
+                    )
+                    utilization.append(
+                        {
+                            "name": name,
+                            "date": date.strftime("%Y-%m-%d"),
+                            "utilization": cu_month["creditUtilization"],
+                        }
+                    )
+            return utilization
+
+        data = self._get_credit_utilization(**kwargs)
+        return _process_utilization(data)
 
     def get_investment_data(self, limit: int = 1000, **kwargs):
         """
@@ -414,3 +652,69 @@ class MintEndpoints(object, metaclass=ABCMeta):
         merged["net"] = merged["assets"] + merged["debts"]
 
         return merged.to_dict("records")
+
+    def get_credit_report_data(
+        self,
+        limit=2,
+        details=True,
+        exclude_inquiries=False,
+        exclude_accounts=False,
+        exclude_utilization=False,
+    ):
+        # Get credit reports. The UI shows 2 by default, but more are available!
+        # At least 8, but could be all the TransUnion reports Mint has
+        # How the "bands" are defined, and other metadata, is available at a
+        # /v1/creditscoreproviders/3 endpoint (3 = TransUnion)
+        credit_report = dict()
+        credit_report["reports"] = self.get_credit_reports(limit=limit)
+
+        # If we want details, request the detailed sub-reports
+        if details:
+            # Get full list of credit inquiries
+            if not exclude_inquiries:
+                credit_report["inquiries"] = self.get_credit_inquiries()
+
+            # Get full list of credit accounts
+            if not exclude_accounts:
+                credit_report["accounts"] = self.get_credit_accounts()
+
+            # Get credit utilization history (~3 months, by account)
+            if not exclude_utilization:
+                credit_report["utilization"] = self.get_credit_utilization()
+
+        return credit_report
+
+    def get_credit_score_data(self):
+        # Request a single credit report, and extract the score
+        report = self.get_credit_report_data(
+            limit=1,
+            details=False,
+            exclude_inquiries=False,
+            exclude_accounts=False,
+            exclude_utilization=False,
+        )
+        try:
+            vendor = report["reports"]["vendorReports"][0]
+            return vendor["creditReportList"][0]["creditScore"]
+        except (KeyError, IndexError):
+            raise Exception("No Credit Score Found")
+
+
+class _ResponseMetadata(object):
+    """
+    Convenience wrapper for pagination
+    """
+
+    def __init__(self, metadata):
+        self.metadata = metadata
+
+    @property
+    def has_next(self):
+        return "link" in self.metadata and any(
+            [i["rel"] == "next" for i in self.metadata["link"]]
+        )
+
+    @property
+    def next_uri_path(self):
+        link = [i for i in self.metadata["link"] if i["rel"] == "next"][0]
+        return link["href"]
