+++ mintapi/rest.py
@@ -5,7 +5,7 @@ Module for REST based API
 import logging
 from typing import Dict, List, Optional, Union
 
-from requests import Response, Session
+from requests import Session
 
 from mintapi.endpoints import MintEndpoints
 
@@ -80,64 +80,6 @@ class RESTClient(MintEndpoints):
     Accessor Methods
     """
 
-    def _paginate(self, data_key: str, metadata_key: str, response: Response, **kwargs):
-        """
-        Mint API appears to use a limit-offset pagination mechanism with
-        href links embedded in responses. Can iterate through sequentially
-        and append results together
-
-        Data schema:
-        {DataType: [], metaData: {}} where DataType is a dynamic key mapping to the endpoint
-        metaData follows a consistent format:
-
-        {
-            'asOf': 'TIMESTAMP',
-            'totalSize': INT,
-            'pageSize': 20,
-            'currentPage': 1,
-            'offset': 0,
-            'limit': 20,
-            'link': [
-                {'otherAttributes': {'method': 'GET'}, 'href': '/v1/accounts', 'rel': 'self'},
-                {'otherAttributes': {}, 'href': '/v1/accounts?offset=20&limit=20', 'rel': 'next'}
-            ]
-        }
-
-        Parameters
-        ----------
-        data_key : str
-            _description_
-        response : Response
-            _description_
-
-        Returns
-        -------
-        _type_
-            _description_
-        """
-        data = []
-
-        json_data = response.json()
-        data.extend(json_data[data_key])
-        metadata = _ResponseMetadata(json_data[metadata_key])
-
-        # drop url params from propagating (href includes the full uri path already)
-        kwargs.pop("params", None)
-
-        while metadata.has_next:
-            response = self.request(
-                uri_path=metadata.next_uri_path,
-                data_key=data_key,
-                metadata_key=metadata_key,
-                paginate=False,
-                **kwargs,
-            )
-            json_data = response.json()
-            data.extend(json_data[data_key])
-            metadata = _ResponseMetadata(json_data[metadata_key])
-
-        return data
-
     def request(
         self,
         *,
@@ -190,29 +132,3 @@ class RESTClient(MintEndpoints):
             )
         else:
             return response
-
-    def get(self, **kwargs):
-        return self.request(method="GET", **kwargs)
-
-    def post(self, **kwargs):
-        return self.request(method="POST", **kwargs)
-
-
-class _ResponseMetadata(object):
-    """
-    Convenience wrapper for pagination
-    """
-
-    def __init__(self, metadata):
-        self.metadata = metadata
-
-    @property
-    def has_next(self):
-        return "link" in self.metadata and any(
-            [i["rel"] == "next" for i in self.metadata["link"]]
-        )
-
-    @property
-    def next_uri_path(self):
-        link = [i for i in self.metadata["link"] if i["rel"] == "next"][0]
-        return link["href"]
